backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
openai==1.12.0
pypdf2==3.0.1
tiktoken==0.6.0
numpy==1.26.4
scikit-learn==1.4.0
python-multipart==0.0.9
Werkzeug==3.0.1
gunicorn==21.2.0